hw6/requirements.txt
coreapi
django==2.0.*
django-cors-headers
djangorestframework
djangorestframework-jwt
Pillow
psycopg2
uwsgi